Create HelloGoodbye
Strings and command-line arguments. Write a program HelloGoodbye.java that  takes two names as command-line arguments and prints hello and goodbye messages as shown below(with the names for the hello message in the same order as the command-line arguments and with the names for the goodbye message in reverse order).