Create Checkerboard
Write a program Checkerboard.java that takes a command-line integer n and plots an n-by-n checkerboard pattern to standard drawing. Color the squares blue and light gray, with the bottom-left square blue. To draw,
Call StdDraw.setScale(0, n) so that x- and y-coordinates of the canvas range from 0 and n.
Call either StdDraw.filledSquare() or StdDraw.filledPolygon() to draw each of the n2 squares.
Make sure that the squares fit snugly in the standard drawing window.
Do not change the canvas size.