Create WorldMap
Write a program WorldMap.java that reads boundary information of a country (or other geographic entity) from standard input and plots the results to standard drawing. A country consists of a set of regions (e.g., states, provinces, or other administrative divisions), each of which is described by a polygon.
Input format.   The first line contains two integers: width and height. The remaining part of the input is divided into regions.

The first entry in each region is the name of the region. For simplicity, names will not contain spaces.
The next entry is an integer specifying the number of vertices in the polygon describing the region.
Finally, the region contains the x- and y-coordinates of the vertices of the polygon.